adds clusterfile
